댓글 프로필 가져오기

+++ src/main/resources/static/assets/js/novel/novelWrite.js
@@ -86,6 +86,19 @@ function getSession(){
 
 //form태그 보내기
 $('.enter').on("click",function () {
+    if(!$('input[name="subNovelTitle"]').val()){
+        alert("소제목을 입력해주세요");
+        return;
+    }
+    if(!$('textarea[name="novelContent"]').val()){
+        alert("작품 내용을 입력해주세요");
+        return;
+    }
+    if(!$('textarea[name="subNovelWriterComment"]').val()){
+        alert("작품 후기를 입력해주세요");
+        return;
+    }
+
     sessionStorage.clear();
     $('input[name="content"]').val($('textarea[name="novelContent"]').val());
     $('input[name="novelNumber"]').val(novelNumber);
@@ -96,6 +109,18 @@ $('.enter').on("click",function () {
 
 //수정 form태그 보내기
 $('.modifyEnter').on("click", function () {
+    if(!$('input[name="modify_sTitle"]').val()){
+        alert("소제목을 입력해주세요");
+        return;
+    }
+    if(!$('textarea[name="modify_novelContent"]').val()){
+        alert("작품 내용을 입력해주세요");
+        return;
+    }
+    if(!$('textarea[name="modify_writerComment"]').val()){
+        alert("작품 후기를 입력해주세요");
+        return;
+    }
     sessionStorage.clear();
     $('input[name="content"]').val($('textarea[name="modify_novelContent"]').val());
     $('input[name="subNovelTitle"]').val($('input[name="modify_sTitle"]').val());
